Refactor intermediate models to prepare for udfs
