+++ packages/op-datasets/src/op_datasets/etl/intermediate/main.py
@@ -1,34 +1,8 @@
-import importlib
-import os
 from op_coreutils.logger import bind_contextvars, clear_contextvars, structlog
 
-from .models import REGISTERED_INTERMEDIATE_MODELS
-from .task import construct_tasks, IntermediateModelsTask
-
-
-def find_modules(path: str):
-    """Finds all python modules that define models in the path.
-
-    This is used to auto-discover models based on convention.
-    """
-    modules = []
-
-    for basename in os.listdir(path):
-        name = os.path.join(path, basename)
-        if os.path.isfile(name) and basename not in ("__init__.py", "registry.py"):
-            modules.append(basename.removesuffix(".py"))
-
-    return sorted(modules)
-
-
-# Python modules under the "models" directory are imported to populate the model registry.
-MODELS_PATH = os.path.join(os.path.dirname(__file__), "models")
-for basename in os.listdir(MODELS_PATH):
-    name = os.path.join(MODELS_PATH, basename)
-    if os.path.isfile(name) and basename not in ("__init__.py", "registry.py"):
-        importlib.import_module(
-            f"op_datasets.etl.intermediate.models.{basename.removesuffix(".py")}"
-        )
+from .registry import REGISTERED_INTERMEDIATE_MODELS
+from .task import IntermediateModelsTask
+from .construct import construct_tasks
 
 
 log = structlog.get_logger()
