--- packages/op-datasets/src/op_datasets/etl/intermediate/models/__init__.py
@@ -1,1 +1,0 @@
-from .registry import REGISTERED_INTERMEDIATE_MODELS
