--- packages/op-datasets/src/op_datasets/etl/intermediate/models/registry.py
@@ -1,5 +1,0 @@
-REGISTERED_INTERMEDIATE_MODELS = {}
-
-
-def register(func):
-    REGISTERED_INTERMEDIATE_MODELS[func.__name__] = func
